requirements.txt
requests==2.31.0
python-dotenv==1.0.0
web3==6.11.3
colorama==0.4.6
schedule==1.2.0
